Cleanup on unused code

+++ TradeHero/TH_ANDROID/src/main/java/com/tradehero/th/fragments/timeline/UserProfileCompactViewHolder.java
@@ -74,7 +74,7 @@ public class UserProfileCompactViewHolder
     {
         if (mEdit != null)
         {
-            if (userProfileDTO != null && userProfileDTO.id == currentUserId.get().intValue())
+            if (userProfileDTO != null && userProfileDTO.id == currentUserId.get())
             {
                 mEdit.setVisibility(View.VISIBLE);
             }
